refactor article page into more files for clarity
